# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Formulator API",
    "version": "0.1.0",
    "description": "REST API contract for the Formulator form builder application.\n\nDerived from the Angular client (`SpreadService`, `SnapService`, `SpillService`), wire DTOs,\nand the mock HTTP interceptor that simulates the backend during local development.\n\n**Spreads** are editable form drafts. **Snaps** are immutable published snapshots\nof a spread's schema, used for response collection. A **spill** is an unique instance of a snap shared for response collection, and contains the response schema.\n\nDeleting a spread does not delete its snaps; affected snaps become orphans with\n`spreadId`, `spreadVersion` and `edition` set to null. Deleting a snap does not affect any of its spills.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development backend"
    }
  ],
...
  "paths": {
    "/spreads": {
      "get": {
        "tags": [
          "Spreads"
        ],
        "operationId": "listSpreads",
        "summary": "List spread metadata",
        "description": "Returns metadata for all spreads without the full page/element tree.\n",
        "responses": {
          "200": {
            "description": "List of spread metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpreadMetaDataDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Spreads"
        ],
        "operationId": "createSpread",
        "summary": "Create a new spread",
        "description": "Creates a new spread with `version` 0, a default schema (one empty page), spread title,\nand current timestamps.\n\nRequest body must provide `spreadTitle`. UUID is server-generated unless `id` is set in body (avoid setting `id` manually except for testing).\n\nReturns **409 Conflict** when `spreadTitle` is already in use or when a client-provided `id` already exists.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpreadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Spread created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpreadDto"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/spreads/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SpreadId"
        }
      ],
      "get": {
        "tags": [
          "Spreads"
        ],
        "operationId": "getSpreadById",
        "summary": "Get spread by ID",
        "description": "Returns the full spread including its schema (title, pages, and elements).",
        "responses": {
          "200": {
            "description": "Spread found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpreadDto"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/SpreadNotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Spreads"
        ],
        "operationId": "updateSpread",
        "summary": "Update spread",
        "description": "Updates a spread by ID. Used for auto-save.\n\nRequired body fields: `id`, `version`, `spreadTitle`, and `schema`.\nThe path `id` and request body `id` must match.\n\nOptimistic concurrency is enforced via the `version` field: the request body\nmust carry the current version. On success the server increments `version`,\nupdates `spreadTitle` and `schema`, and refreshes `lastModifiedAt`.\n\nReturns **409 Conflict** for a version mismatch or when `spreadTitle` is\nalready used by another spread.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpreadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spread updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpreadDto"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/SpreadNotFound"
          },
          "409": {
            "description": "Version conflict or duplicate spread title",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/VersionConflictDetail"
                    },
                    {
                      "$ref": "#/components/schemas/ProblemDetail"
                    }
                  ]
                },
                "examples": {
                  "versionConflict": {
                    "value": {
                      "status": 409,
                      "title": "Version conflict",
                      "detail": "Expected version 2 but found 3.",
                      "code": "VERSION_CONFLICT",
                      "expectedVersion": 2,
                      "actualVersion": 3
                    }
                  },
                  "duplicateSpreadTitle": {
                    "value": {
                      "status": 409,
                      "title": "Conflict",
                      "detail": "duplicate spread title.",
                      "code": "CONFLICT"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Spreads"
        ],
        "operationId": "deleteSpread",
        "summary": "Delete spread",
        "description": "Deletes a spread by ID. Does not delete associated snaps; published snaps\nno longer reference spread.\n",
        "responses": {
          "204": {
            "description": "Spread deleted"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/SpreadNotFound"
          }
        }
      }
    },
    "/snaps": {
      "get": {
        "tags": [
          "Snaps"
        ],
        "operationId": "listSnaps",
        "summary": "List snap metadata",
        "description": "Returns metadata for snaps. When `spreadId` is provided, returns only snaps\npublished from that spread. When omitted, returns all snaps.\n\nThe `title` field is denormalized from `schema.title` on the full snap.\n\nOrphan editions (source spread deleted) are included when listing all snaps\nand have `spreadId: null`, `spreadVersion: null` and `edition: null`.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/SpreadIdQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "List of snap metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SnapMetaDataDto"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      },
      "post": {
        "tags": [
          "Snaps"
        ],
        "operationId": "createSnap",
        "summary": "Publish a snap",
        "description": "Publishes an immutable snapshot of a spread's current schema.\n\n- Deep-clones the spread's `schema` at publish time (spread row unchanged).\n- Sets `schema.title` on the snap from `snapTitle`.\n- Records `spreadVersion` from the spread's current `version`.\n- Assigns `edition` as the highest existing edition for that spread plus\n  one (starting at 1).\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSnapRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Snap published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapDto"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/SpreadNotFound"
          }
        }
      }
    },
    "/snaps/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SnapId"
        }
      ],
      "get": {
        "tags": [
          "Snaps"
        ],
        "operationId": "getSnapById",
        "summary": "Get snap by ID",
        "description": "Returns the full snap including its immutable schema snapshot.",
        "responses": {
          "200": {
            "description": "Snap found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapDto"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/SnapNotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Snaps"
        ],
        "operationId": "deleteSnap",
        "summary": "Delete snap",
        "description": "Removes the snap from the builder. After a successful delete, list and get\nrequests for that snap return 404 \u2014 a hard delete from the client's perspective.\n\nThe server retains the snap record internally (soft-close) so existing spills\ncan still resolve the published schema for response data.\n",
        "responses": {
          "204": {
            "description": "Snap removed from the client; no longer returned by list or get"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/SnapNotFound"
          }
        }
      }
    },
    "/spills": {
      "get": {
        "tags": [
          "Spills"
        ],
        "operationId": "listSpills",
        "summary": "List spill metadata",
        "description": "Returns metadata for all spills matching the provided `snapId`, omitting the response schema.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/SnapIdQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "List of spill metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpillMetaDataDto"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      },
      "post": {
        "tags": [
          "Spills"
        ],
        "operationId": "createSpill",
        "summary": "Share a spill",
        "description": "Creates a unique response collection record for a responder.\n\n- Request body must provide `snapId`, `email` (responder's email) and valid `rSchema`.\n- Optionally, `firstName` and `lastName` (of responder) can be provided.\n- Form `schema` is not copied onto the spill; it is resolved from `snapId` when needed.\n- `lastModifiedAt` and `completedAt` are null until the spill is submitted (`PATCH /spills/{id}`).\n- Spill `id` acts as the unique token when sharing the spill instance with responder.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpillRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Spill created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpillMetaDataDto"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/SnapNotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "SpreadNotFound": {
        "description": "Spread not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "spreadNotFound": {
                "value": {
                  "status": 404,
                  "title": "Spread not found",
                  "detail": "No spread exists with id 00000000-0000-0000-0000-000000000000.",
                  "code": "SPREAD_NOT_FOUND"
                }
              }
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict \u2014 e.g. duplicate spread title or duplicate client-provided id",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "duplicateSpreadTitle": {
                "value": {
                  "status": 409,
                  "title": "Conflict",
                  "detail": "duplicate spread title.",
                  "code": "CONFLICT"
                }
              },
              "duplicateId": {
                "value": {
                  "status": 409,
                  "title": "Conflict",
                  "detail": "duplicate id",
                  "code": "CONFLICT"
                }
              }
            }
          }
        }
      },
...
    },
...
    "parameters": {
      "SpreadId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Spread UUID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "schemas": {
      "SpreadMetaDataDto": {
        "type": "object",
        "required": [
          "id",
          "spreadTitle",
          "createdAt",
          "lastModifiedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "spreadTitle": {
            "type": "string",
            "minLength": 1,
            "description": "Display title. Leading/trailing whitespace is trimmed; empty-after-trim is rejected."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateSpreadRequest": {
        "type": "object",
        "required": [
          "spreadTitle"
        ],
        "additionalProperties": false,
        "properties": {
          "spreadTitle": {
            "type": "string",
            "description": "Display title."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional client-provided spread ID. When omitted, the server generates a UUID."
          }
        }
      },
...
      "SpreadDto": {
        "type": "object",
        "required": [
          "id",
          "version",
          "spreadTitle",
          "schema",
          "createdAt",
          "lastModifiedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "spreadTitle": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "minimum": 0,
            "description": "Optimistic locking counter; incremented on each successful update."
          },
          "schema": {
            "$ref": "#/components/schemas/Schema"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  }
}